<folder name(Header.css, App.js, SearchBar.js): changing of the folder RealHeade and Header name>                                                                                                  <changing of the name of the folders for a more accurate>
