+++ src/views/Homepage/index.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import RandomRecipe from '../../components/RandomRecipe/index.js'
 import RecipeCategories from '../../components/RecipeCategories'
 // import AreaCategories from '../../components/AreaCategories'
-import SearchBar from '../../components/Header/index'
+import SearchBar from '../../components/SearchBar/index'
